修改 GetSysKernalVersion() 的类型和返回值

+++ extern.hpp
@@ -4,11 +4,11 @@
 #include<fstream>
 #include<ShlObj.h>
 #include<windows.h>
-unsigned long GetSysKernalVersion(){
+DWORD GetSysKernalVersion(){
     typedef void(__stdcall*NTPROC)(DWORD*,DWORD*,DWORD*);
     HINSTANCE inst{LoadLibrary(TEXT("ntdll.dll"))};
     NTPROC getSysKernalVersion{(NTPROC)GetProcAddress(inst,"RtlGetNtVersionNumbers")};
     DWORD major,minor;
     getSysKernalVersion(&major,&minor,NULL);
-    return major*10+minor;
+    return major*10U+minor;
 }
